src/test/java/RepoTests/ExpenseRepoTests/ExpenseCategoryRepoTests/TravelRepoTests/BusRepoImpTest.java
package RepoTests.ExpenseRepoTests.ExpenseCategoryRepoTests.TravelRepoTests;

import BudgetBuddi.Domain.Expenses.ExpsenseCategories.Travel.Bus;
import BudgetBuddi.Repository.ExpenseRepositories.ExpenseCategoryRepositories.TravelRepositories.BusRepository;
import BudgetBuddi.Repository.Implentation.ExpensesRepoImplementation.ExpenseCategoryReposImplementaion.TravelRepoImplementation.BusRepoImp;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import java.util.ArrayList;

public class BusRepoImpTest {

    private BusRepository repository;

    @Before
    public void setUp() throws Exception
    {
        this.repository = BusRepoImp.getRepository();
    }

    @Test
    public void getRepository() {
    }

    @Test
    public void getAll()
    {
        ArrayList<Bus> buses = this.repository.getAll();
        Assert.assertEquals(1, buses.size());
    }

    @Test
    public void create()
    {
        this.repository.create(null);
        Assert.assertEquals(null, null);
    }

    @Test
    public void update() {
    }

    @Test
    public void delete()
    {
        this.repository.delete(null);
        Assert.assertEquals(null, null);
    }

    @Test
    public void read() {
    }
}